requirements.txt
flask
pyyaml
blinker
konfig
Flask-WebTest
flake8
pytest-cov
autopep8